added the strategy pattern and added the description the current pizza that is cooking

+++ src/payment/Payment.java
@@ -1,5 +1,5 @@
 package payment;
 
 public interface Payment {
-    void pay(int amount);
+    void pay(double amount);
 }
